improve: Various fixes to contracts and off-chain infra based on e2e test findings

+++ packages/core/contracts/insured-bridge/BridgePool.sol
@@ -93,7 +93,7 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, Lockable {
 
     // Data from L2 deposit transaction.
     struct DepositData {
-        uint8 chainId;
+        uint256 chainId;
         uint64 depositId;
         address payable l1Recipient;
         address l2Sender;
@@ -368,7 +368,7 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, Lockable {
      *      quoteTimestamp. The OO acts to verify the correctness of this realized fee. Can not exceed 50%.
      */
     function relayDeposit(
-        uint8 chainId,
+        uint256 chainId,
         uint64 depositId,
         address payable l1Recipient,
         address l2Sender,
